Move definitions to sub-module, only define for H503 for now
